Refactoring into a much more suitable file structure
