+++ lsa/record/__init__.py
@@ -1,0 +1,3 @@
+from .bib import BibtexRecordParser, BibtexRecordIterator
+from .froac import FroacRecordParser, FroacRecordIterator
+from .isi import IsiRecordParser, IsiRecordIterator
