+++ lsa/record/froac.py
@@ -1,0 +1,56 @@
+import hashlib
+
+from xml.dom import minidom
+
+from .base import RecordParser
+from .base import RecordIterator
+
+from .util import gen_to_list
+from .util import xml_to_text
+
+
+class FroacRecordParser(RecordParser):
+
+    def clear(self, field, raw):
+        if field in ['title', 'keywords', 'uuid', 'description']:
+            return getattr(self, '_clear_' + field)(raw)
+        raise NotImplementedError(
+            'The field ' + field + 'has not been implemented')
+
+    def parse(self, raw):
+        if isinstance(raw, str):
+            return super().parse(minidom.parseString(raw))
+        return super().parse(raw)
+
+    @xml_to_text
+    def _clear_title(self, raw):
+        return raw.getElementsByTagName('lom:title').item(0)
+
+    @xml_to_text
+    def _clear_description(self, raw):
+        return raw.getElementsByTagName('lom:description').item(0)
+
+    @gen_to_list
+    def _clear_keywords(self, raw):
+        keywords = raw.getElementsByTagName('lom:keyword')
+        for keyword in keywords:
+            yield keyword.firstChild.nodeValue
+
+    def _clear_uuid(self, raw):
+        sha = hashlib.sha1()
+        sha.update(raw.toxml().encode('utf-8'))
+        return sha.hexdigest()
+
+
+class FroacRecordIterator(RecordIterator):
+
+    '''
+    Iterates plain txt froac records in a file.
+    '''
+
+    parser_class = FroacRecordParser
+
+    def get_buffer(self):
+        dom = minidom.parse(self.filename)
+        for dom_element in dom.getElementsByTagName('record'):
+            yield dom_element
